distinguished between database files for mongoose dbms and postgres dbms. moved schema for postgres db to separate folder, and wrote new index.js to connect db and drafted initial model function for get request. in process of writing seed reviews file.
