+++ database-pg/seedReviews.js
@@ -1,0 +1,2 @@
+const { Pool } = require('pg');
+const connection = require('./index.js');
