+++ server/index.js
@@ -5,7 +5,7 @@ const cors = require('cors');
 const app = express();
 const path = require('path');
 
-const dbHelpers = require('../database/dbHelpers.js');
+const dbHelpers = require('../database-mg/dbHelpers.js');
 
 // Serve Static files
 app.use(express.static(path.join(__dirname, '../client/dist')));
